+++ middlewar-controller/src/main/java/com/middlewar/controllers/buildings/ItemFactoryController.java
@@ -36,16 +36,16 @@ public class ItemFactoryController {
 
     @RequestMapping(value = "/me/base/{baseId}/factory/{factoryId}/module/create/{id}", method = RequestMethod.POST)
     public Response createModule(@AuthenticationPrincipal Account pAccount,
-                                 @PathVariable(value = "baseId") Long baseId,
-                                 @PathVariable(value = "factoryId") Long factoryId,
+                                 @PathVariable(value = "baseId") int baseId,
+                                 @PathVariable(value = "factoryId") int factoryId,
                                  @PathVariable(value = "id") String itemId) {
         return controllerManagerWrapper.wrap(() -> factoryManager.createModule(baseManager.getOwnedBase(baseId, playerManager.getCurrentPlayerForAccount(pAccount)), factoryId, itemId));
     }
 
     @RequestMapping(value = "/me/base/{baseId}/factory/{factoryId}/structure/create/{id}", method = RequestMethod.POST)
     public Response createStructure(@AuthenticationPrincipal Account pAccount,
-                                    @PathVariable(value = "baseId") Long baseId,
-                                    @PathVariable(value = "factoryId") Long factoryId,
+                                    @PathVariable(value = "baseId") int baseId,
+                                    @PathVariable(value = "factoryId") int factoryId,
                                     @PathVariable(value = "id") String itemId) {
         return controllerManagerWrapper.wrap(() -> factoryManager.createStructure(baseManager.getOwnedBase(baseId, playerManager.getCurrentPlayerForAccount(pAccount)), factoryId, itemId));
     }
